// src/app/features/grid/ag-grid-extention/enums/column-type.enum.ts
export enum ColumnType {
  Autocomplete = 'autocomplete',
  Checkbox = 'checkbox',
  Date = 'date',
  DateTime = 'dateTime',
  Dropdown = 'dropdown',
  Hidden = 'hidden',
  Id = 'id',
  Number = 'number',
  NumberDecimal = 'numberDecimal',
  NumberInt = 'numberInt',
  Status = 'status',
  Text = 'text'
}